Refactoring to fix imports of package brancheslist, adding BranchesListFragment and fixing the interface problem from fragments implementations. All View interfaces from fragments was removed.
