+++ app/src/main/java/br/com/luisfernandez/github/client/koin/ModelModule.kt
@@ -1,9 +1,9 @@
 package br.com.luisfernandez.github.client.koin
 
+import br.com.luisfernandez.github.client.repodetails.brancheslist.BranchesListModelImpl
+import br.com.luisfernandez.github.client.repodetails.brancheslist.BranchesListModel
 import br.com.luisfernandez.github.client.repodetails.commitslist.CommitsListModel
 import br.com.luisfernandez.github.client.repodetails.commitslist.CommitsListModelImpl
-import br.com.luisfernandez.github.client.repodetails.contributorslist.BranchesListModel
-import br.com.luisfernandez.github.client.repodetails.contributorslist.BranchesListModelImpl
 import br.com.luisfernandez.github.client.repodetails.contributorslist.ContributorsListModel
 import br.com.luisfernandez.github.client.repodetails.contributorslist.ContributorsListModelImpl
 import br.com.luisfernandez.github.client.repodetails.issuelist.IssueListModel
