+++ app/src/main/java/br/com/luisfernandez/github/client/repodetails/brancheslist/BranchesListModelImpl.kt
@@ -1,4 +1,4 @@
-package br.com.luisfernandez.github.client.repodetails.contributorslist
+package br.com.luisfernandez.github.client.repodetails.brancheslist
 
 import br.com.luisfernandez.github.client.http.GitHubService
 import br.com.luisfernandez.github.client.pojo.BranchResponse
