+++ app/src/main/java/br/com/luisfernandez/github/client/repodetails/contributorslist/ContributorsListFragment.kt
@@ -14,16 +14,15 @@ import br.com.luisfernandez.github.client.extensions.setVisible
 import br.com.luisfernandez.github.client.http.model.GitHubErrorBody
 import br.com.luisfernandez.github.client.http.model.ServerError
 import br.com.luisfernandez.github.client.pojo.ContributorResponse
-import br.com.luisfernandez.github.client.pojo.IssueResponse
 import br.com.luisfernandez.github.client.repodetails.RepoDetailsPagerAdapter.Companion.OWNER
 import br.com.luisfernandez.github.client.repodetails.RepoDetailsPagerAdapter.Companion.REPONAME
-import kotlinx.android.synthetic.main.fragment_issues_list.*
+import kotlinx.android.synthetic.main.fragment_contributors_list.*
 import kotlinx.android.synthetic.main.view_state_empty.*
 import kotlinx.android.synthetic.main.view_state_error.*
 import kotlinx.android.synthetic.main.view_state_loading.*
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class ContributorsListFragment : Fragment(), ContributorsListView {
+class ContributorsListFragment : Fragment() {
 
     lateinit var owner: String
     lateinit var repoName: String
@@ -65,7 +64,7 @@ class ContributorsListFragment : Fragment(), ContributorsListView {
         })
     }
 
-    override fun handleError(serverError: ServerError<GitHubErrorBody>) {
+    private fun handleError(serverError: ServerError<GitHubErrorBody>) {
         this.showErrorState()
 
         if (serverError.errorBody != null) {
@@ -87,21 +86,21 @@ class ContributorsListFragment : Fragment(), ContributorsListView {
         recyclerView.setGone()
     }
 
-    override fun showLoading() {
+    private fun showLoading() {
         layoutProgress.setVisible()
         layoutEmpty.setGone()
         layoutError.setGone()
         recyclerView.setGone()
     }
 
-    override fun showEmpty() {
+    private fun showEmpty() {
         layoutProgress.setGone()
         layoutEmpty.setVisible()
         layoutError.setGone()
         recyclerView.setGone()
     }
 
-    override fun showContent(content: List<ContributorResponse>) {
+    private fun showContent(content: List<ContributorResponse>) {
         if (content.isEmpty()) {
             showEmpty()
         } else {
